Refactor to shift forecast to new  RCForecasterBase class; remove ReservoirComputerBase.
