+++ src/orc/models/__init__.py
@@ -1,6 +1,6 @@
 """Implementations of common model architectures."""
 
 from orc.models import esn
-from orc.models.esn import ESN, train_ESN_forecaster
+from orc.models.esn import ESNForecaster, train_ESNForecaster
 
-__all__ = ["esn", "ESN", "train_ESN_forecaster"]
+__all__ = ["esn", "ESNForecaster", "train_ESNForecaster"]
